Data_Generation_Preparation/3d_adjust/requirements.txt
PyYAML==5.4.1
natsort==7.1.0
opencv-python==4.5.1.48
svgwrite==1.4.1
tqdm==4.56.0
matplotlib==3.3.4
parse==1.19.0
scikit-spatial==5.2.0
scipy==1.5.4
